Created level select scene. Added Volume slider. Fixed some bugs.
Added a Level select scene that is access before playing the game. This
replaces the dropdown menu that was in the Options menu. Now the Options
menu just has the volume controls. Also the a custom sprite for the
slider was added for the volume control. Some bugs were fixed with audio
that caused the audio to bounce audio to stop playing after an object
was matched.